requirements.txt
playwright
psutil
curl_cffi
cryptography
Pillow
beautifulsoup4
discord.py
numpy
google-api-python-client
google-auth-oauthlib
redis
python-dotenv